cleanup and refactors, remove high-level GraphGridSession class.
